Split authentication app tests into separate files
