+++ src/thunderbird_accounts/authentication/tests/test_utils.py
@@ -1,0 +1,115 @@
+from unittest.mock import MagicMock, patch
+
+import freezegun
+from django.conf import settings
+from django.contrib import messages
+from django.contrib.admin import AdminSite
+from django.core.exceptions import PermissionDenied
+from django.forms import model_to_dict
+from django.http import HttpRequest
+from urllib.parse import quote
+from django.test import Client as RequestClient, override_settings
+from django.test import TestCase
+from django.utils.translation import gettext_lazy as _
+
+from requests import Response
+
+from thunderbird_accounts.authentication.admin import CustomUserAdmin
+from thunderbird_accounts.authentication.clients import RequestMethods
+from thunderbird_accounts.authentication.exceptions import ImportUserError
+from thunderbird_accounts.authentication.middleware import AccountsOIDCBackend
+from thunderbird_accounts.authentication.models import User
+from thunderbird_accounts.authentication.reserved import is_reserved, servers, support
+from thunderbird_accounts.mail.models import Account, Email
+from thunderbird_accounts.core.tests.utils import build_keycloak_success_response
+from thunderbird_accounts.authentication.models import AllowListEntry
+
+class IsReservedUnitTests(TestCase):
+    def test_brand_names(self):
+        for brand in ['thunderbird', 'mozilla', 'firefox', 'help', 'support', 'mzla']:
+            self.assertTrue(is_reserved(brand))
+
+    def test_brand_plus_variants(self):
+        # (brands)? allows more than one match
+        for name in ['thunderbirdthunderbird', 'supportsupport']:
+            self.assertFalse(is_reserved(name))
+
+    def test_brand_support_help_suffix_patterns(self):
+        # ^(brand).*support?$ and customer_support and help
+        for name in [
+            'thunderbirdpro_customer_support',
+            'thunderbirdpro_support',
+            'thunderbird_customer_support',
+            'thunderbird_support',
+            'thunderbird-support',
+            'mzla_support',
+            'mzla_help',
+            'mozilla_support',
+            'firefox_help',
+        ]:
+            self.assertTrue(is_reserved(name))
+
+        # ^(brand).*email?$ and ^(brand).*org?$
+        for name in ['thunderbird_email', 'firefox_org']:
+            self.assertTrue(is_reserved(name), name)
+
+    def test_official_and_real_variants(self):
+        for name in [
+            'official_thunderbird',
+            'officialsupport',
+            'realmozilla',
+            'firefox_real',
+            'support_official',
+            'mozilla_real',
+        ]:
+            self.assertTrue(is_reserved(name))
+
+    def test_mzla_test_variants(self):
+        for name in ['mzla-test', 'mzla-test.123', 'mzla-test.alpha.beta']:
+            self.assertTrue(is_reserved(name))
+
+    def test_common_example_usernames(self):
+        for name in ['username', 'user_name', 'user', 'exampleuser', 'example_name', 'example-user', 'test']:
+            self.assertTrue(is_reserved(name))
+
+    def test_servers(self):
+        for name in ['admin', 'root', 'webmaster', 'postmaster', 'superuser', 'administrator']:
+            self.assertTrue(is_reserved(name))
+
+    def test_team_and_contact(self):
+        for name in [
+            'team',
+            'hr',
+            'accounts_team',
+            'engineering',
+            'engineering_team',
+            'marketing_team',
+            'design',
+            'design_team',
+            'contactus',
+            'contact_us',
+        ]:
+            self.assertTrue(is_reserved(name))
+
+    def test_internal_names(self):
+        # select a few to hardcode test
+        for name in ['root', 'postmaster', 'support', 'marketing']:
+            self.assertTrue(is_reserved(name), name)
+
+        for name in servers + support:
+            self.assertTrue(is_reserved(name), name)
+
+    def test_birbs(self):
+        for name in ['roc', 'ezio', 'mithu', 'ava', 'callum', 'sora', 'robin', 'nemo']:
+            self.assertTrue(is_reserved(name))
+
+    def test_non_reserved(self):
+        reserved_names_related = ['mozillafan', 'supporter', 'helper', 'hostmastery', 'contacts']
+        unrelated = ['randomuser', 'this_should_not_be_a_problem', '123_asdf', 'asdf_123', '123asdf', 'asdf123']
+        for name in reserved_names_related + unrelated:
+            self.assertFalse(is_reserved(name))
+
+    def test_partial_matches_should_pass(self):
+        # Anchors ^...$ mean full-string match only
+        for name in ['user123', 'myusernamex', 'rooted', 'teamwork', 'contacting']:
+            self.assertFalse(is_reserved(name))
